feat: Introduce `forbidden()`
### What?
Similar to the notFound() function and the `notFound.js` file, this PR follows the convention and introduces the `forbidden()` function along side the `forbiddden.js` file.

### Why?
While the available `notFound` and `redirect` functions can get you a long way, there is only so much you can do with them.

Developers need a way to handle the authorized state of a user across the multiple "contexts" next has, by that I am referring to components, api routes, and server actions.